requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
azure-ai-formrecognizer==3.3.0
python-multipart==0.0.6
azure-search-documents==11.5.2
openai==1.12.0
numpy==1.26.4
pydantic==2.11.3
pydantic-core==2.33.1
typing-extensions>=4.9.0
httpx==0.27.2